moving utils to utils
