+++ lib/raster.js
@@ -3,14 +3,11 @@ var mapnik = require('mapnik');
 var srs = require('srs');
 var invalid = require('./invalid');
 var path = require('path');
+var utils = require('./utils');
 
 mapnik.register_default_input_plugins();
 
 module.exports = Raster;
-module.exports.convertToMeters = convertToMeters;
-module.exports.getSpatialResolutions = getSpatialResolutions;
-module.exports.getUnitType = getUnitType;
-module.exports.getValidSpatialResolutions = getValidSpatialResolutions;
 
 function Raster(filepath) {
   var geotransform;
@@ -162,65 +159,14 @@ Raster.prototype.getLayers = function(callback) {
  * @param srs
  * @returns unit type (String)
  */
-function getUnitType(srs) {
-  var possibleUnits = ['m', 'ft', 'mi', 'km', 'us-ft', 'us-mi'];
-  var i;
-
-  for (i = 0; i < possibleUnits.length; i++) {
-    if (srs.indexOf('+units=' + possibleUnits[i]) !== -1) return possibleUnits[i];
-  }
-
-  if (srs.indexOf('+units=') === -1 && srs.indexOf('+proj=longlat') !== -1) return 'decimal degrees';
-
-  //Default to meters for now, if nothing matches
-  else return 'm';
-}
-
-function convertToMeters(pixelSize, unit) {
-  var circumference = 40075000;
-  var conversions = {
-    m: function(x) { return x; },
-
-    ft: function(x) { return x / 0.3048; },
-
-    mi: function(x) { return x / 1609.34; },
-
-    km: function(x) { return x / 1000; },
-
-    'us-ft': function(x) { return x / 0.3048; },
-
-    'us-mi': function(x) { return x / 1609.34; },
-
-    'decimal degrees': function(x) { return x / 360 * circumference; }
-  };
-
-  var x = conversions[unit](pixelSize[0]);
-  var y = conversions[unit](pixelSize[1]);
-
-  return [x, y];
-}
-
-function getSpatialResolutions() {
-  var circumference = 40075000;
-  var zoomLevels = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19];
-  return zoomLevels.map(function(z) {
-    return circumference * Math.cos(0) / Math.pow(2, (z + 8));
-  });
-}
-
-function getValidSpatialResolutions(spatialResolutions, pixelSize) {
-  return spatialResolutions.filter(function(res) {
-    return res > pixelSize;
-  });
-}
 
 Raster.prototype.getZooms = function(callback) {
-  var pixelSize = convertToMeters(this.details.pixelSize, getUnitType(this.projection));
+  var pixelSize = utils.convertToMeters(this.details.pixelSize, utils.getUnitType(this.projection));
 
-  var spatialResolutions = getSpatialResolutions();
+  var spatialResolutions = utils.getSpatialResolutions();
 
   var _this = this;
-  var validSpatialResolutions = getValidSpatialResolutions(spatialResolutions, _this.details.pixelSize[0]);
+  var validSpatialResolutions = utils.getValidSpatialResolutions(spatialResolutions, _this.details.pixelSize[0]);
 
   this.details.pixelSize = pixelSize;
   return callback(
